+++ packages/jest-shared-config/jest-preset.js
@@ -1,0 +1,9 @@
+/** @type {import('jest').Config} */
+const config = {
+  preset: "ts-jest",
+  transform: {
+    "^.+\\.ts$": "ts-jest",
+  },
+};
+
+module.exports = config;