+++ packages/server/src/internalHandlers/currency.ts
@@ -1,16 +1,33 @@
-import { CurrencyList, schemaCurrencyList } from "@ledgerhq/wallet-api-core";
+import {
+  Currency,
+  CurrencyList,
+  schemaCurrencyList,
+} from "@ledgerhq/wallet-api-core";
 import { firstValueFrom } from "rxjs";
 import type { RPCHandler } from "../types";
+
+function filterCurrenciesByCurrencyIds(
+  currencies: Currency[],
+  currencyIds: string[]
+) {
+  const currencyIdsSet = new Set(currencyIds);
+
+  return currencies.filter((currency) => currencyIdsSet.has(currency.id));
+}
 
 export const list: RPCHandler<CurrencyList["result"]> = async (
   req,
   context
 ) => {
-  schemaCurrencyList.params.parse(req.params);
+  const safeParams = schemaCurrencyList.params.parse(req.params);
 
-  const currencies = await firstValueFrom(context.currencies$);
+  const { currencyIds } = safeParams;
+
+  const allCurrencies = await firstValueFrom(context.currencies$);
 
   return {
-    currencies,
+    currencies: currencyIds
+      ? filterCurrenciesByCurrencyIds(allCurrencies, currencyIds)
+      : allCurrencies,
   };
 };