+++ packages/server/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ledgerhq/wallet-api-server
 
+## 0.7.1
+
+### Patch Changes
+
+- [#45](https://github.com/LedgerHQ/wallet-api/pull/45) [`b63f820`](https://github.com/LedgerHQ/wallet-api/commit/b63f820722caf18a236105bdc2e80ec4f73935a3) Thanks [@Justkant](https://github.com/Justkant)! - fix: apduHex params naming
+
+## 0.7.0
+
+### Minor Changes
+
+- [#41](https://github.com/LedgerHQ/wallet-api/pull/41) [`4572761`](https://github.com/LedgerHQ/wallet-api/commit/45727617dd415589f39fe7fcd6474a4ec681b0ce) Thanks [@Justkant](https://github.com/Justkant)! - feat: device.transport support
+
+### Patch Changes
+
+- Updated dependencies [[`4572761`](https://github.com/LedgerHQ/wallet-api/commit/45727617dd415589f39fe7fcd6474a4ec681b0ce)]:
+  - @ledgerhq/wallet-api-core@0.7.0
+
+## 0.6.0
+
+### Minor Changes
+
+- [#37](https://github.com/LedgerHQ/wallet-api/pull/37) [`016d399`](https://github.com/LedgerHQ/wallet-api/commit/016d399d9e1daa98eae30dc1050c4c5bf44980c6) Thanks [@IAmMorrow](https://github.com/IAmMorrow)! - fixed errors; added permissions
+
+### Patch Changes
+
+- Updated dependencies [[`016d399`](https://github.com/LedgerHQ/wallet-api/commit/016d399d9e1daa98eae30dc1050c4c5bf44980c6)]:
+  - @ledgerhq/wallet-api-core@0.6.0
+
 ## 0.5.4
 
 ### Patch Changes